Remove duplicate upload code
This commit refactors cloudinit and ignition codepaths to
utilize more of the same code for uploading volumes.
